requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
streamlit-sortables>=0.2.0
xlsxwriter>=3.1.0
